pyproject.toml
[project]
name = "license-plate"
version = "0.1.0"
description = "Synthetic license plate data generation, plate & character detection training, and a lightweight inference package for end-to-end LPR."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "albumentations>=2.0.8",
    "google-genai>=1.45.0",
    "httpx>=0.28.1",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pydantic>=2.12.2",
    "rfdetr>=1.3.0",
    "roboflow>=1.2.11",
    "supervision>=0.26.1",
    "torch>=2.9.0",
]
